+++ cmd/frontend/graphqlbackend/search_results.go
@@ -520,98 +520,20 @@ func (r *searchResolver) evaluateLeaf(ctx context.Context) (_ *SearchResultsReso
 	return rr, err
 }
 
-type searchResultDeduper struct {
-	seenFileMatches   map[string]*FileMatchResolver
-	seenRepoMatches   map[string]*RepositoryResolver
-	seenCommitMatches map[string]*CommitSearchResultResolver
-	seenDiffMatches   map[string]*CommitSearchResultResolver
-}
-
-func (d *searchResultDeduper) init() {
-	if d.seenFileMatches == nil {
-		d.seenFileMatches = make(map[string]*FileMatchResolver)
-		d.seenRepoMatches = make(map[string]*RepositoryResolver)
-		d.seenCommitMatches = make(map[string]*CommitSearchResultResolver)
-		d.seenDiffMatches = make(map[string]*CommitSearchResultResolver)
-	}
-}
-
-// Add adds a SearchResultResolver to the deduper, merging it into
-// a previously added SearchResultResolver if the URL has already been seen
-func (d *searchResultDeduper) Add(r SearchResultResolver) {
-	d.init()
-
-	if fileMatch, ok := r.ToFileMatch(); ok {
-		if prev, seen := d.seenFileMatches[fileMatch.uri]; seen {
-			prev.appendMatches(fileMatch)
-			return
-		}
-		d.seenFileMatches[fileMatch.uri] = fileMatch
-		return
-	}
-
-	if repoMatch, ok := r.ToRepository(); ok {
-		if _, seen := d.seenRepoMatches[repoMatch.URL()]; seen {
-			return
-		}
-		d.seenRepoMatches[repoMatch.URL()] = repoMatch
-		return
-	}
-
-	if commitMatch, ok := r.ToCommitSearchResult(); ok {
-		if commitMatch.DiffPreview() != nil {
-			if _, seen := d.seenDiffMatches[commitMatch.URL()]; seen {
-				return
-			}
-			d.seenDiffMatches[commitMatch.URL()] = commitMatch
-			return
-		} else {
-			if _, seen := d.seenCommitMatches[commitMatch.URL()]; seen {
-				return
-			}
-			d.seenCommitMatches[commitMatch.URL()] = commitMatch
-			return
-		}
-	}
-
-	return
-}
-
-// Deduped returns a slice of SearchResultResolvers, deduplicated from
-// the SearchResultResolvers that were added with Add
-func (d *searchResultDeduper) Deduped() []SearchResultResolver {
-	total := len(d.seenFileMatches) + len(d.seenRepoMatches) + len(d.seenCommitMatches) + len(d.seenDiffMatches)
-	r := make([]SearchResultResolver, 0, total)
-	for _, v := range d.seenFileMatches {
-		r = append(r, v)
-	}
-	for _, v := range d.seenRepoMatches {
-		r = append(r, v)
-	}
-	for _, v := range d.seenCommitMatches {
-		r = append(r, v)
-	}
-	for _, v := range d.seenDiffMatches {
-		r = append(r, v)
-	}
-	return r
-}
-
 // unionMerge performs a merge of file match results, merging line matches when
 // they occur in the same file.
 func unionMerge(left, right *SearchResultsResolver) *SearchResultsResolver {
-	var dedup searchResultDeduper
+	dedup := NewDeduper()
 
-	// Register all the left results in the dedupper
+	// Add results to maps for deduping
 	for _, leftResult := range left.SearchResults {
 		dedup.Add(leftResult)
 	}
-
 	for _, rightResult := range right.SearchResults {
 		dedup.Add(rightResult)
 	}
 
-	left.SearchResults = dedup.Deduped()
+	left.SearchResults = dedup.Dedupe()
 	left.Stats.Update(&right.Stats)
 	return left
 }
